+++ jaxlib/mosaic/dialect/tpu/tpu_ops.cc
@@ -1058,17 +1058,13 @@ LogicalResult MatmulOp::verify() {
     auto dimension_numbers = getDimensionNumbers().value();
     auto lhs_contracting_dims = dimension_numbers.getLhsContractingDims();
     auto rhs_contracting_dims = dimension_numbers.getRhsContractingDims();
-    if (lhs_contracting_dims.size() != 1) {
-      emitOpError("Not implemented: lhs contracting dims must be of size 1");
+    if (lhs_contracting_dims.size() != rhs_contracting_dims.size() ||
+        lhs_contracting_dims.empty()) {
+      emitOpError(
+          "Not implemented: number of lhs and rhs contracting dims must be "
+          "equal and non-zero");
       return failure();
     }
-    if (rhs_contracting_dims.size() != 1) {
-      emitOpError("Not implemented: rhs contracting dims must be of size 1");
-      return failure();
-    }
-
-    auto lhs_contracting_dim = lhs_contracting_dims[0];
-    auto rhs_contracting_dim = rhs_contracting_dims[0];
 
     auto lhs_batch_dims = dimension_numbers.getLhsBatchDims();
     auto rhs_batch_dims = dimension_numbers.getRhsBatchDims();
@@ -1104,12 +1100,14 @@ LogicalResult MatmulOp::verify() {
       return failure();
     }
 
-    if (lhs_ty.getShape()[lhs_contracting_dim] !=
-        rhs_ty.getShape()[rhs_contracting_dim]) {
-      emitOpError(
-          "Not implemented: lhs and rhs contracting dims must be of the same "
-          "size");
-      return failure();
+    for (int64_t i = 0; i < lhs_contracting_dims.size(); ++i) {
+      if (lhs_ty.getDimSize(lhs_contracting_dims[i]) !=
+          rhs_ty.getDimSize(rhs_contracting_dims[i])) {
+        emitOpError(
+            "Not implemented: lhs and rhs contracting dims must be of the same "
+            "size");
+        return failure();
+      }
     }
 
     if (lhs_batch_dims.size() != rhs_batch_dims.size()) {